Added: -c attack -getUserIndex method -getItemIndex method -checkUserExistance method -checkItemExistance method -new inputVerification standard for verifying user input -health to user creation
All of this implements the attack mechanic in it's simplest form.
Created items can now be used to attack other users. Item damage will be
calculated and the users health in their database entry will be changed.

WHAT THIS BUILD IS MISSING:
-Completion messages
-Specific error messages if player or item does not exist
-Case error fixes. As in, eveything needs to be converted to lower case
-Item deletion**
-Cross check with Garrett

DO NOT MERGE INTO MAIN WITHOUT COMPLETING THE ABOVE